+++ linkoauth/__init__.py
@@ -19,135 +19,6 @@
 # the Initial Developer. All Rights Reserved.
 #
 # Contributor(s):
-#
-import abc
 
-from webob.exc import HTTPRedirection
-from services.pluginreg import PluginRegistry
-
-from linkoauth import facebook_, google_, twitter_, yahoo_, linkedin_
-from linkoauth.sstatus import ServicesStatus
-from linkoauth.errors import BackendError, DomainNotRegisteredError
-
-#from linkoauth.live_ import LiveResponder
-#from linkoauth.openidconsumer import OpenIDResponder
-
-__all__ = ['Responder', 'get_responder', 'Requester', 'get_requester',
-           'Services']
-
-
-class Responder(PluginRegistry):
-    """Abstract Base Class for the responder APIs."""
-    plugin_type = 'responder'
-
-    @abc.abstractmethod
-    def get_name(self):
-        """Returns the name of the plugin."""
-
-    @abc.abstractmethod
-    def request_access(self, request, url, session):
-        """Returns a redirect url
-
-        Args:
-            request: WebOb request object
-            url: Routes URL generator callable
-            session: Session object
-
-        Returns:
-            return url
-        """
-
-# pre-register provided backends
-Responder.register(twitter_.TwitterResponder)
-Responder.register(facebook_.FacebookResponder)
-Responder.register(google_.GoogleResponder)
-Responder.register(yahoo_.YahooResponder)
-Responder.register(linkedin_.responder)
-
-
-def get_responder(domain, **kw):
-    try:
-        return Responder.get(domain, **kw)
-    except KeyError:
-        raise DomainNotRegisteredError(domain)
-
-
-class Requester(PluginRegistry):
-    """Abstract Base Class for the requester APIs."""
-    plugin_type = 'requester'
-
-    @abc.abstractmethod
-    def get_name(self):
-        """Returns the name of the plugin."""
-
-    @abc.abstractmethod
-    def sendmessage(self, message, options={}):
-        """xxx"""
-
-    @abc.abstractmethod
-    def getcontacts(self, start=0, page=25, group=None):
-        """xxx"""
-
-
-# pre-register provided backends
-Requester.register(twitter_.TwitterRequester)
-Requester.register(facebook_.FacebookRequester)
-Requester.register(google_.GoogleRequester)
-Requester.register(yahoo_.YahooRequester)
-Requester.register(linkedin_.api)
-
-
-def get_requester(domain, account, **kw):
-    try:
-        return Requester.get(domain, account=account, **kw)
-    except KeyError:
-        raise DomainNotRegisteredError(domain)
-
-
-# high-level
-class Services(ServicesStatus):
-
-    def __init__(self, services, servers=None, ttl=600):
-        requesters = [req.get_name() for req in Requester._abc_registry]
-        responders = [res.get_name() for res in Responder._abc_registry]
-
-        for service in services:
-            if service not in requesters:
-                raise DomainNotRegisteredError(service)
-
-            if service not in responders:
-                raise DomainNotRegisteredError(service)
-
-        ServicesStatus.__init__(self, services, servers, ttl)
-
-    def _updated(func):
-        def __updated(self, domain, *args, **kw):
-            domain = str(domain)
-            try:
-                res = func(self, domain, *args, **kw)
-            except BackendError, e:
-                self.update_status(domain, False)
-                return None, e.args[0]
-            except HTTPRedirection:
-                self.update_status(domain, True)
-                raise
-            else:
-                if len(res) == 2 and res[0] is not None:
-                    self.update_status(domain, True)
-            return res
-        return __updated
-
-    @_updated
-    def sendmessage(self, domain, account, message, options={}, **kw):
-        return get_requester(domain, account).sendmessage(message, options,
-                                                          **kw)
-
-    @_updated
-    def getcontacts(self, domain, account, start=0, page=25, group=None, **kw):
-        return get_requester(domain, account, **kw).getcontacts(start, page,
-                                                                group)
-
-    @_updated
-    def request_access(self, domain, request, url, session, **kw):
-        return get_responder(domain, **kw).request_access(request, url,
-                                                          session)
+# exposing the class registers existing backends
+from linkoauth.backends import Services   # NOQA
