+++ linkoauth/backends/facebook_.py
@@ -30,7 +30,7 @@ import random
 import copy
 import logging
 
-from linkoauth.base import OAuth2
+from linkoauth.oauth import OAuth2
 from linkoauth.errors import OAuthKeysException
 from linkoauth.util import build_url, config
 from linkoauth.protocap import HttpRequestor
