+++ linkoauth/backends/google_.py
@@ -60,7 +60,7 @@ from linkoauth.oid_extensions import OAuthRequest
 from linkoauth.oid_extensions import UIRequest
 from linkoauth.openidconsumer import ax_attributes, attributes
 from linkoauth.openidconsumer import OpenIDResponder
-from linkoauth.base import get_oauth_config
+from linkoauth.oauth import get_oauth_config
 from linkoauth.protocap import ProtocolCapturingBase, OAuth2Requestor
 from linkoauth.errors import BackendError, OptionError, OAuthKeysException
